Refactor code in order to separate functionality
